+++ swing/common-ui/src/main/java/is/codion/swing/common/ui/component/table/FilterTableConditionPanel.java
@@ -18,8 +18,6 @@
  */
 package is.codion.swing.common.ui.component.table;
 
-import is.codion.common.event.Event;
-import is.codion.common.event.EventObserver;
 import is.codion.common.model.table.TableConditionModel;
 import is.codion.swing.common.ui.Utilities;
 import is.codion.swing.common.ui.component.table.ColumnConditionPanel.ConditionState;
@@ -28,7 +26,7 @@ import java.awt.BorderLayout;
 import java.util.ArrayList;
 import java.util.Collection;
 import java.util.Map;
-import java.util.Optional;
+import java.util.function.Consumer;
 import java.util.stream.Collectors;
 
 import static is.codion.swing.common.ui.component.table.FilterTableColumnComponentPanel.filterTableColumnComponentPanel;
@@ -40,23 +38,25 @@ import static java.util.stream.Collectors.toMap;
 /**
  * A default filter table condition panel.
  * @param <C> the column identifier type
- * @see #filterTableConditionPanel(TableConditionModel, Collection, FilterTableColumnModel)
+ * @see #filterTableConditionPanel(TableConditionModel, Collection, FilterTableColumnModel, Consumer)
  */
 public final class FilterTableConditionPanel<C> extends TableConditionPanel<C> {
 
 	private final Collection<ColumnConditionPanel<C, ?>> conditionPanels;
 	private final FilterTableColumnModel<C> columnModel;
-	private final Event<?> initializedEvent = Event.event();
+	private final Consumer<TableConditionPanel<C>> onPanelInitialized;
 
 	private FilterTableColumnComponentPanel<C> componentPanel;
 	private boolean initialized;
 
 	private FilterTableConditionPanel(TableConditionModel<C> conditionModel,
 																		Collection<ColumnConditionPanel<C, ?>> conditionPanels,
-																		FilterTableColumnModel<C> columnModel) {
+																		FilterTableColumnModel<C> columnModel,
+																		Consumer<TableConditionPanel<C>> onPanelInitialized) {
 		super(conditionModel);
 		this.conditionPanels = unmodifiableList(new ArrayList<>(requireNonNull(conditionPanels)));
 		this.columnModel = requireNonNull(columnModel);
+		this.onPanelInitialized = onPanelInitialized == null ? panel -> {} : onPanelInitialized;
 	}
 
 	@Override
@@ -77,22 +77,19 @@ public final class FilterTableConditionPanel<C> extends TableConditionPanel<C> {
 						.collect(Collectors.toList());
 	}
 
-	@Override
-	public Optional<EventObserver<?>> initializedEvent() {
-		return Optional.of(initializedEvent.observer());
-	}
-
 	/**
 	 * @param <C> the column identifier type
 	 * @param conditionModel the condition model
 	 * @param conditionPanels the condition panels
 	 * @param columnModel the column model
+	 * @param onPanelInitialized called when the panel has been initialized
 	 * @return a new {@link FilterTableConditionPanel}
 	 */
 	public static <C> FilterTableConditionPanel<C> filterTableConditionPanel(TableConditionModel<C> conditionModel,
 																																					 Collection<ColumnConditionPanel<C, ?>> conditionPanels,
-																																					 FilterTableColumnModel<C> columnModel) {
-		return new FilterTableConditionPanel<>(conditionModel, conditionPanels, columnModel);
+																																					 FilterTableColumnModel<C> columnModel,
+																																					 Consumer<TableConditionPanel<C>> onPanelInitialized) {
+		return new FilterTableConditionPanel<>(conditionModel, conditionPanels, columnModel, onPanelInitialized);
 	}
 
 	@Override
@@ -115,12 +112,12 @@ public final class FilterTableConditionPanel<C> extends TableConditionPanel<C> {
 
 	private void initialize() {
 		if (!initialized) {
+			initialized = true;
 			setLayout(new BorderLayout());
 			Map<C, ColumnConditionPanel<C, ?>> conditionPanelMap = conditionPanels.stream()
 							.collect(toMap(panel -> panel.conditionModel().identifier(), identity()));
 			componentPanel = filterTableColumnComponentPanel(columnModel, conditionPanelMap);
-			initialized = true;
-			initializedEvent.run();
+			onPanelInitialized.accept(this);
 		}
 	}
 }